requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.2.2
langchain-text-splitters>=0.2.2
langchain-core>=0.2.0
langchain-groq>=0.1.3
pypdf>=4.2.0
numpy>=1.24.0
tiktoken>=0.6.0
python-dotenv>=1.0.1
